feat(keyboard): transfer led update info from master to slave
To be able to synchronize LEDs between halves we need to have a
single source of truth. Master sends current state of the keyboard
to slave, which will use this information to render LEDs.
Current limitation is that the other half still uses it's own
LED configuration, so it won't be correct if both sides have different
firmware. To solve this we could maybe try to send configuration
via serial and use it instead of the one that has been flashed, but
this is quite complicated, and let's better think some time if there
is a simpler solution.