fix: avoid double stack usage in init by using &'static in resources
All the resources in Shared/Local are allocated 2 times due to how
rtic works - once in main() (generated code) and once in init().
This can be avoided by using MaybeUninit and writing directly into
that memory (but this adds a level of indirection which potentially
decreases performance).

The following resources have been replaced:
* usb: 1656 B (will be trimmed down in the future)
* leds: 884 B (now led_controller)
* keyboard: 1156 B